+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add new npm script `components:publish-cdn` to published files to Cellar
 * Add manual CSS to style not yet defined components (will be reworked and automated later)
 * Add new npm script `components:build-cdn:versions-list` to update the list of available versions (published on CDN)
+* Add Jenkinsfile to automate CDN publication on new tag
 
 ## 5.3.1 (2020-11-16)